Resolve swiftlint warnings
