+++ Shared/Sources/Shared/Services/ReviewPromptController.swift
@@ -31,8 +31,7 @@ public enum ReviewPromptController {
 
         // Don't request more than once per 120 days
         if let lastRequest = lastRequestTime,
-           Date().timeIntervalSince(lastRequest) < 120 * 24 * 60 * 60
-        {
+           Date().timeIntervalSince(lastRequest) < 120 * 24 * 60 * 60 {
             return
         }
